Nuove Modifiche

+++ src/main/java/com/websystique/springmvc/dao/AlbumDao.java
@@ -1,5 +1,6 @@
 package com.websystique.springmvc.dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.websystique.springmvc.model.Album;
@@ -8,12 +9,12 @@ public interface AlbumDao {
 
 	List<Album> findByMetaTag(String meta);
 	
-	Album findById(int id);
+	Album findById(BigInteger id);
 	
 	void save(Album album);
 	
-	List<Album> findAllByUserId(int userId);
+	List<Album> findAllByUserId(BigInteger userId);
 	
-	void deleteById(int id);
+	void deleteById(BigInteger id);
 	
 }
