+++ src/main/java/com/websystique/springmvc/dao/UserDaoImpl.java
@@ -1,5 +1,6 @@
 package com.websystique.springmvc.dao;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.hibernate.Criteria;
@@ -17,7 +18,7 @@ import com.websystique.springmvc.model.User;
 public class UserDaoImpl extends AbstractDao<Integer, User> implements UserDao {
 
 	
-	public User findById(int id) {
+	public User findById(BigInteger id) {
 		User user = getByKey(id);
 		if(user!=null){
 			Hibernate.initialize(user.getId());
