+++ src/main/java/com/websystique/springmvc/service/AlbumService.java
@@ -1,5 +1,6 @@
 package com.websystique.springmvc.service;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import com.websystique.springmvc.model.Album;
@@ -9,11 +10,11 @@ public interface AlbumService {
 
 	List<Album> findByMetaTag(String meta);
 	
-	Album findById(int id);
+	Album findById(BigInteger id);
 	
-	List<Album> findAllByUserId(int userId);
+	List<Album> findAllByUserId(BigInteger userId);
 	
-	void deleteById(int id);
+	void deleteById(BigInteger id);
 	
 	void saveAlbum(Album document);
